Hi Charlie,
Can you make the following changes to the release calendar:
•	change the time requirement for standard pre-production to 4 business hours
•	change the time requirement for standard production to 8 business hours
•	change the verbiage in the form accordingly - for the 2 above
•	add the following description for urgent pre-production/production releases below the verbiage for "Pre-production deployments...":
"Urgent pre-production and production releases do not have a required notification period. However, urgent releases require approval from either or both the Development Manager or/and IT Manager before creation and both must be added to the People and Roles section as SW Dev Lead and IT Lead, respectively. If approval is not granted prior to the creation of an urgent item, the item will NOT be processed until approval is gained."
•	change the e-mail reminder default to 8 hours before start time
•	add QA as a Source for Production deployments, but keep INT as the default

Thanks!
- John