Correct the SyncSvc/Web.config to reference a service which implements SyncService.SyncSvc.
This is a TEMPORARY FIX.  The correct one is to enforce some consistency around the service contract's name.

==============

The class which implements the Synchronization service is named SyncSvc.  Ah, but what’s its full name?

In SyncSvc.svc.cs, the fully qualified name is NCI.EmergencyAlert.SyncSvc.

Likewise, in the portion of the service web.config which makes the service available, the contract is also advertised as NCI.EmergencyAlert.SyncSvc



But when it’s a client?  That’s where it gets confusing.

The app.config for the reference in EmergencyAlertLibrary calls it NCI.EmergencyAlert.SyncSvc.

The service contains a reference to a remote instance of the service, it’s looking for SyncService.SyncSvc (The client portion of the web.config actually said Fred.SyncSvc, which was apparently copied from the client configuration portion of a previous version of the the test application's config file.)

And then the test application is looking for a contract called SyncSvc.SyncSvc.


Bottom line is, they should all be consistent about referring to the contract as NCI.EmergencyAlert.SyncSvc.