So this error is slightly unrelated to the original TIR (https://tracker.nci.nih.gov/browse/OCEDRUINTR-927). Sridhar when evaluating the pager for search, did notice the paging was almost consistently off by one page less. For example, there were supposed to be 4 pages of results, but only 3 showed...
This was due to a math error. When dividing the total number of results by the pager, often times we were left with a float. Like, 3.32 would be 3 pages with a remainder, so technically a fourth page of results should exist. However, the $pager functionality rounded down. I added a ceiling to the division ceil($total/$limit) to get the rounded up value instead.