User/Org Automation - https://tracker.nci.nih.gov/browse/OCEDRUINTR-582 Updated the nightly imports to include more logging functionality. I created my own basic logging class to handle writing from a custom database table as we did not want to clutter up watchdog. (It should handle the reading/display as well, but didn't have time to add that in - I just reworked what watchdog was doing)
Updated the NEDUser migration class to create authmap entires when users are imported into the system. Currently, this goes through all of the active users and creates it if it is not already there. This could probably be updated in order to be more efficient. 


TODO: Still need to import body text for organizations (this should be a new migration class).
TODO: Still need to configure the cron jobs to run for importing nightly.