Got rid of the entity reference error and fixed the FMI linking error.
The entity ref error was solved by moving workbench_moderation_save below node_save. This was not seen locally as technically this area had been commented out in the script.

The second error is because something was entering an if statement that should've been an elseif statement and getting super confused. (The map was also wrong... it was $mappedOrgs instead of $mappedContent)