https://tracker.nci.nih.gov/browse/OCEDRUINTR-1406 - Created Content Tags and Content Owner pages.
Content Owners is incomplete, but stable.  Major issues with development of story involved the multi-value Contact and Owner fields - These fields cannot be sorted unless broken into a separate column for each value, but also cannot be sorted if using anything other than one row per value.  This results in multiple lines for a piece of content when multiple values exist, which will need to be corrected by a not-exactly-trivial query alter.

Also, the default method or sorting is not by the formatted value displayed, but the underlying username or term id.  So, the query will also have to be altered to take the displayed value into account, or perhaps the user or ogranization name can be displayed instead via a relation, so that sorting naturally works.

Also ran into issues getting exposed filtering on these fields to work, due to any entered value being checked against the username or tid.  A working autocomplete implementation should actually help, as the resulting value from the filter should also be the underlying id of the related field.

Finally, having to break out contact and owner into two columns each makes the view very wide, and difficult to fit into the standard view width.  Currently, these two fields have all of the values compiled into a single row, which at least removes duplicates and cuts down on view width.  It'd be worth reviewing this story when possible to see if any features are required or if an alternate approach could save time.