After chasing a 508 issue with reCAPTCHA around a lot, it basically comes down to we have to do a little hacking. The reCAPTCHA module pretty much is just a dump from Google.
Google/reCAPTCHA module do support a Custom themes, but they are completely barren: http://www.google.com/recaptcha/demo/custom
The Custom theme is the only one that supports all of the pretty themeing:
https://developers.google.com/recaptcha/docs/customization

Even with trying to re-hack Google's code before it goes out, it comes out with a hackey Custom theme with a pretty themed square below it. The image is above the pretty square, and really the whole thing is a mess.

insertBefore should get to the DOM, which is what 508 standards need. This should comply, but may not be the best way. Officially recommending we theme are own, but this will do for now.