The reason slots were not updating is because the "slot within a slot" (when the render was being handled solely by node--*.tpl and not going thru hook_preprocess_page()!) Thus, this had a different method of checking child slot items. It turned out it was related to do the the $current_vid of what was being loaded.
The old code had this:
   ----> if((count($url) > 2 ? $url[2] == 'draft' : 0) && $url[1] == $node->nid)
As the code to determine if we should load the draft version ID or not. This was done via stmp_slots_template_check_rel_removed(), as the second parameter here was always $current_vid. However, the $url being passed in was obviously not the correct $nid when it came to a slot within a slot. For example, say: A (nid:1) had B (nid: 2) as a child and B had a child, that was C (nid: 3). When viewing DRAFT of A, we should see draft of A&B&C. However, the URL at the page had nid 1, not nid 3 like this code was expecting.

I'm not sure why it was there in the first place, but is not consistent across the slots (e.g. preprocess page does not have this and neither does carousel items) PLUS everything looks fine so I think it's safe to presume it's ok to remove.

Code was checked regarding the following types:
- Carousel items (already worked, in theory)
- Slot lists (? Not sure about this one...)
- Highlighted content

Code was not fixed regarding the "Slot Image" as I can't figure out how to check that on the front end.

https://tracker.nci.nih.gov/browse/OCEDRUPAL-380
https://tracker.nci.nih.gov/browse/OCEDRUPAL-398