https://tracker.nci.nih.gov/browse/OCEDRUPAL-402 https://tracker.nci.nih.gov/browse/OCEDRUPAL-401 https://tracker.nci.nih.gov/browse/OCEDRUPAL-400 https://tracker.nci.nih.gov/browse/OCEDRUPAL-398
Modified the terms of _standardtemplates_get_child_item(). Originally, I was assuming that $node->vid always corresponded to the published node. However, this does not really seem to be true. Thus, I added logic to make sure I was pulling the PUBLISHED node and comparing it to the node we are VIEWING.