***WORKING COPY***
fixed the issue with pager randomly jumping ahead by like 100s of pages. the problem was in the pager.tpl; sometimes (currentPage + 1) read currentPage as a String instead of an Int so that
(currentPage + 1) where currentPage=3 evaluated to 31 instead of 4
This was fixed with a parseInt()

Also fixed the confusion with the inherent .reset() being confused with the created function .resetResults().

Still needs cleanup, but this is a good copy.