http://oceworks.nci.nih.gov/CTB/Lists/Issues/DispForm.aspx?ID=2924 https://tracker.nci.nih.gov/browse/OCEOTHER-20
Added a new page for "Tips" and removed the previous popup for "Review our Tips" on the homepage. Of course this included a rewiring of the "Review our tips" link that I am still not completely satisfied with.

Page was named "Tips", but Meghan getting in touch with client about this. Also about the possibility of a jump list.

Two different resource files were used to get the translation/HTML looking clean. Code behind was copied from "es_help", "en_help", and "home" though, so that part may not be perfect.