https://tracker.nci.nih.gov/browse/OCEDRUINTR-1377
Updating the about_this_page_standard mini panel so that it no longer uses field_event_type. In order to achieve this, I had to create Custom Content that includes <h3> title HTML & a token to display the name. I also had to set the visibility rules on this item such that it didn't show up on other content types, e.g. a Form would have a section in the About This Page in the left column that said:

*Event Type*
Form