Commit for https://tracker.nci.nih.gov/browse/OCEDRUINTR-1598
This was to fix the wrong OFFICE OF THE DIRECTOR ISSUE.

To elaborate on what was happening...
The JSON came through as "OFFICE OF THE DIRECTOR". This was then converted, somehow, magically, to a term ID in order for it to be stored in field_nci_organization. However, in its magical conversion, it was just matching whatever the last OFFICE OF THE DIRECTOR was in the nci orgs vocabulary... which there is a lot of.

First step to fix was to distinguish what was being selected versus what was being sent. Because we have a fancy div and spans to show the Org breadcrumbs, we had to use JSON. Since we had to use JSON, we were only transmitting what was selected by a string. Since we were only transmitting a string, we had to append a tid to distinguish the OD.

After appending the tid, we got our step 2. The changed "selected text" was rejected. We needed to figure out how to force its acceptance. Summarizing a lot of work, it's validation was done by taxonomy_autocomplete_validate(). This validation function OPENLY states that it DOES NOT take tids with the string it tries to match. This turns out to be where the heart of our problem was. We override this validation function with a custom one by tweaking the form api on the widget element. We accept it if we can pull out a proper tid (thus distinguishing the term name), do nothing if there is nothing sent (not as duh as it sounds), and then set an error if they DO send something, but we can't load a term out of it.

There was a third step. The default behavior is just to load the name of the field for autocomplete of tags. Thus, after reloading any edited documented, you would now see "OFFICE OF THE DIRECTOR" instead of the proper/previously selected "OFFICE OF THE DIRECTOR (16)". THUS we had to rewrite how the text got set in this box. THis is where the custom #value_callback came in. In which the tid is re-attached to the textual value.