Brought hypercare1.0 branch changes into trunk.
Unfortunately, svn merge --reintegrate didn't work, presumably because trunk and the branch were kept in sync using techniques other than svn merge. So I created fresh exports from trunk and branch, generated a patch, and applied it to an up-to-date sandbox for trunk. This commit is the result of that process.