Fix code to detect when user_set_authmaps() must be invoked.
The original test in NCIDirectoryImporter::UsersUpdate() was not
adequate, as it compared the sso_login_name field of the user
account's object with the NIHSSOUSERNAME value from the NED.
The problem with this approach was that when a user's account
was dropped from the NED feed temporarily (for example, because
the organization was changed from NCI to some other value),
the row in the authmap table was dropped, but the sso_login_name
field was left intact. Then when the account was restored to the
feed the authmap would never get populated for the user, because
the field still matched the NIHSSOUSERNAME value received from
the NED. That test was replaced by an actual check of the authmap
table. (JIRA ticket OCEDRUINTR-1639)